Add render string code
This was a trip. Off-by-one errors galore. But, it was worth it. The
editor displays both UTF-8 and long lines just about perfectly. The
only potential issue is emoji width, but IIRC that depends on the
terminal anyway. So, final version, I'll put a define into config.h or
something.

Some quality of life keys (including the page keys but also EOF and
BOF) and this'll be about as good a file viewer as less, lol.