Add basic multiple buffer support
In terms of new code it's a gimme. An important feature supported by
all "proper" text editors; and essential for multiple windows.

In terms of restructuring, editorConfig has been removed from most of
the code except for region.h (where we need it to set or use the kill)
and two Unicode functions in row.h and undo.h. It'd be really pog to
remove those and not to have to pass around the world. If I did that,
I could restrict the usage just to main! Of course, that'd be
premature, as I'd like to move as much as possible *out* of main and
I'm not sure what, that I'd like to move out, would need that data.

The next step will be to implement C-x C-f and friends.