Fix off-by-one line-wrapping issues
This was actually a two-fold problem, which required an interesting
solution:

1. Clearing the lines was erasing the end of the line, for some
reason (perhaps the cursor not going to the next line automagically,
terminals are weird). So, replace clearing the lines one-at-a-time
with clearing the whole screen at the start of the rendering code.
2. There needs to be a separate line for the cursor to go on at the
end of long lines, so put one in explicitly on the edge case.

Love the fast turnaround in C. This program compiles in a flash on
Linux!

Closes #3