Use ranges to define word boundaries
Most of ASCII outside of the alphanumerics and some punctuation is a
boundary of some kind; use ranges to express this. This is a big
improvement; I think the usual expectation is it's only alphanums and
a small handful of punctuation that counts as word characters. This
shouldn't give us any performance degradation either, as the number of
expressions has actually decreased (11->9).